refactor: move review brief support contracts to output crate
