+++ crates/cli/src/bin/schema_emit.rs
@@ -28,7 +28,6 @@ use serde_json::{Map, Value};
 use fallow_cli::audit_brief::{
     DiffTriage, GraphFacts, ReviewBriefOutput, ReviewBriefSchemaVersion, ReviewEffort, RiskClass,
 };
-use fallow_cli::audit_focus::{ConfidenceFlag, FocusLabel, FocusMap, FocusScore, FocusUnit};
 use fallow_cli::audit_walkthrough::{WalkthroughGuide, WalkthroughValidation};
 use fallow_cli::health_types::{
     ComplexityViolation, ContributorEntry, ContributorIdentifierFormat, CoverageGapSummary,
@@ -80,9 +79,10 @@ use fallow_core::duplicates::{
     RefactoringKind, RefactoringSuggestion,
 };
 use fallow_output::{
-    AcceptedJudgment, AgentSchema, ChangeAnchor, Decision, DecisionAction, DecisionActionType,
-    DecisionCategory, DecisionSurface, DecisionSurfaceOutput, DecisionSurfaceSchemaVersion,
-    DecisionWithActions, DirectionUnit, RejectedJudgment, ReviewDirection, TruncationNote,
+    AcceptedJudgment, AgentSchema, ChangeAnchor, ConfidenceFlag, Decision, DecisionAction,
+    DecisionActionType, DecisionCategory, DecisionSurface, DecisionSurfaceOutput,
+    DecisionSurfaceSchemaVersion, DecisionWithActions, DirectionUnit, FocusLabel, FocusMap,
+    FocusScore, FocusUnit, RejectedJudgment, ReviewDirection, TruncationNote,
 };
 use fallow_output::{
     CloneFamilyAction, CloneFamilyActionType, CloneGroupAction, CloneGroupActionType,
